Improved reliability for reading data from APS 9-BM, 20-BM and 20-ID.
The data file from these beamlines includes names for the columns in
the header, which xafs_beamlines.py tries to read. In some cases the
assumptions made about the format of these column names were invalid,
such as whether the column name includes a '*' character and what its
significance is.

This commit uses a regular expression to fully describe (hopefully)
the layout of these column names, and sorts them into the correct
ordered list.