small change to athena project writing for consistency with Athena
If a Larch group is created with a name, i.e.
   g=Group(name='my data')
then the group's name will be used as the label when exporting that
group to an Athena project file.  Because the name, if provided,
presumably carries semantic information about the data, it is a better
choice for the Athena label than Athena's silly 5-letter "hash key",
which is was never intended to be used in a semantic context.

If the Larch group is not explicitly named, then the label will
default to the 5-letter hash key, which is the current behavior.